Move User's code into /src for cleaner structure
