+++ User/src/UserController.js
@@ -1,4 +1,4 @@
-const userService = require('#UserService.js');
+const userService = require('#src/UserService.js');
 
 const login = async (req, res, next) => {
   try {
